# docs/api/whatsapp-ai-openapi.yaml
openapi: 3.0.3
info:
  title: autolumiku WhatsApp AI API
  description: API for WhatsApp AI Automation - customer chat & staff commands management
  version: 1.0.0
  contact:
    name: autolumiku Support
    email: support@autolumiku.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.autolumiku.com/v1
    description: Production server
  - url: https://staging-api.autolumiku.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /whatsapp-ai/status:
    get:
      tags:
        - Status
      summary: Get WhatsApp connection status
      description: Check if WhatsApp is connected and get basic stats
      operationId: getWhatsAppStatus
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          example: "8dd6398e-b2d2-4724-858f-ef9cfe6cd5ed"
      responses:
        '200':
          description: Status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/WhatsAppStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /whatsapp-ai/stats:
    get:
      tags:
        - Analytics
      summary: Get conversation statistics
      description: Retrieve detailed stats for conversations and messages
      operationId: getStats
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ConversationStats'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    WhatsAppStatus:
      type: object
      properties:
        isConnected:
          type: boolean
          example: true
        phoneNumber:
          type: string
          example: "+628123456789"
        clientId:
          type: string
          example: "tenant_8dd6398e_1732789123456"
        lastConnectedAt:
          type: string
          format: date-time
        totalConversations:
          type: integer
          example: 45
        activeConversations:
          type: integer
          example: 12
        todayMessages:
          type: integer
          example: 128
        aiResponseRate:
          type: integer
          description: Percentage (0-100)
          example: 85
...
    ConversationStats:
      type: object
      properties:
        total:
          type: integer
        active:
          type: integer
        escalated:
          type: integer
        customerChats:
          type: integer
        staffCommands:
          type: integer
        avgResponseTime:
          type: number
          description: In seconds
        aiAccuracy:
          type: number
          description: Percentage (0-100)
...
  responses:
    BadRequest:
      description: Bad request - validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
...